update spec to match reality
While the documentation (and therefore the spec) specify a two letter
(ISO 3166) country code, the value returned for the example in the docs
does not conform. The discrepancy reveals that the field does not have
adequate restrictions on the backend. I removed the restriction on the
spec (although I did not change the wording) until we can fix the
backend.