change prism errors default to false
By default, don't blow up a run when we get an error. That lets
us test failure conditions as well as success conditions. We're still
validating request and response envelopes regardless as that is a
different option within Prism. Needing to do this has to do with
the implementation of throwing and catching errors with Promises in
tape (and node-tape, for that matter). At a later date we can probably
change this setting, but as it doesn't affect us we don't need to.