fix mailtype issue
All form factors had wording that applied only to letters(!)

The way composition and inheritance work in OpenAPI v3.0, it isn't
possible to change properties of an attribute, only add properties
to the attribute. To support different options and/or wording for
mailtype, we either need to leave it out of the included model
(what I choose in this commit) or switch to full on polymorphism.
Which blows.