use enums with address discriminators
Although it may seem somewhat counterintuitive, best practice for
specifying a constant field is to use an enum with a single option.

The reason the single-valued enum option works best is that the
tools find it easiest (and fastest) to interpret. Constant time
ftw.

A different rule says that the property used as a discriminator
must be a required field. Of course, a field with a default need
not be required as it will be present in any case.

Building on all of the above, we bring you a more robust and much
faster version of address polymorphism.